Add Spring Doc / Swagger

+++ cambio-service/src/main/java/com/joaoeduardo/cambioservice/model/ConvertedCambioDTO.java
@@ -2,5 +2,5 @@ package com.joaoeduardo.cambioservice.model;
 
 import java.math.BigDecimal;
 
-public record ConvertedCambioDTO(String from, String to, Double conversionFactor, Double convertedValue, String enviroment) {
+public record ConvertedCambioDTO(String from, String to, Double conversionFactor, BigDecimal convertedValue, String enviroment) {
 }
